pyproject.toml
[project]
name = "cardscanner"
version = "0.1.0"
requires-python = ">=3.9"
dependencies = [
  "opencv-python",
  "numpy",
  "Pillow",
  "python-dotenv",
  "pytesseract",
  "openai>=1.0.0",
]

[project.scripts]
cardscanner = "cli:main"
